+++ lib/l10n/app_localizations/app_localizations_es.g.dart
@@ -20,7 +20,33 @@ class AppLocalizationsEs extends AppLocalizations {
   }
 
   @override
-  String get error_error => 'Error';
+  String get error_snack_bar => 'Error:';
+
+  @override
+  String get error_widget_title => 'An error has occurred.';
+
+  @override
+  String get error_widget_description =>
+      'Please report this issue on GitHub or via mail. Attach a screenshot of this page and the logs that you can copy or export below. As a precaution, you should also export your notes.';
+
+  @override
+  String get error_widget_disabled_secure_flag =>
+      'The setting to flag the app as secure is disabled until the next restart to enable screenshots.';
+
+  @override
+  String get error_widget_button_export_notes => 'Export notes';
+
+  @override
+  String get error_widget_button_copy_logs => 'Copy logs';
+
+  @override
+  String get error_widget_button_export_logs => 'Export logs';
+
+  @override
+  String get error_widget_button_create_github_issue => 'Create GitHub issue';
+
+  @override
+  String get error_widget_button_send_mail => 'Send mail';
 
   @override
   String get navigation_notes => 'Notas';
@@ -32,19 +58,19 @@ class AppLocalizationsEs extends AppLocalizations {
   String get navigation_settings => 'Ajustes';
 
   @override
-  String get navigation_settings_appearance => 'Appearance';
+  String get navigation_settings_appearance => 'Apariencia';
 
   @override
-  String get navigation_settings_behavior => 'Behavior';
+  String get navigation_settings_behavior => 'Comportamiento';
 
   @override
   String get navigation_settings_editor => 'Editor';
 
   @override
-  String get navigation_settings_backup => 'Backup';
+  String get navigation_settings_backup => 'Copia de seguridad';
 
   @override
-  String get navigation_settings_about => 'About';
+  String get navigation_settings_about => 'Acerca de';
 
   @override
   String get button_ok => 'Aceptar';
@@ -53,25 +79,25 @@ class AppLocalizationsEs extends AppLocalizations {
   String get button_cancel => 'Cancelar';
 
   @override
-  String get button_sort_title => 'Title';
+  String get button_sort_title => 'Título';
 
   @override
-  String get button_sort_ascending => 'Ascending';
+  String get button_sort_ascending => 'Ascendente';
 
   @override
   String get settings_appearance => 'Apariencia';
 
   @override
-  String get settings_appearance_description => 'Language, theme, text scaling, notes tiles';
+  String get settings_appearance_description => 'Idioma, tema, tamaño del texto, fichas de notas';
 
   @override
-  String get settings_appearance_application => 'Application';
+  String get settings_appearance_application => 'Aplicación';
 
   @override
   String get settings_language => 'Idioma';
 
   @override
-  String get settings_language_contribute => 'Contribute';
+  String get settings_language_contribute => 'Contribuir';
 
   @override
   String get settings_theme => 'Tema';
@@ -89,7 +115,7 @@ class AppLocalizationsEs extends AppLocalizations {
   String get settings_dynamic_theming => 'Tema dinámico';
 
   @override
-  String get settings_dynamic_theming_description => 'Generate colors from the system';
+  String get settings_dynamic_theming_description => 'Generar colores desde el sistema';
 
   @override
   String get settings_black_theming => 'Negro puro';
@@ -98,161 +124,162 @@ class AppLocalizationsEs extends AppLocalizations {
   String get settings_black_theming_description => 'Fondo negro puro para el tema oscuro';
 
   @override
-  String get settings_text_scaling => 'Text scaling';
+  String get settings_text_scaling => 'Escala del texto';
 
   @override
-  String get settings_appearance_notes_tiles => 'Notes tiles';
+  String get settings_appearance_notes_tiles => 'Fichas de las notas';
 
   @override
-  String get settings_show_titles_only => 'Titles only';
+  String get settings_show_titles_only => 'Sólo títulos';
 
   @override
-  String get settings_show_titles_only_description => 'Only show the titles of the notes';
+  String get settings_show_titles_only_description => 'Sólo mostrar los títulos de las notas';
 
   @override
-  String get settings_show_titles_only_disable_in_search_view => 'Disable titles only in search view';
+  String get settings_show_titles_only_disable_in_search_view => 'Deshabilitar mostrar solo los títulos en la búsqueda';
 
   @override
   String get settings_show_titles_only_disable_in_search_view_description =>
-      'Disable the option to only show the titles when in the search view';
+      'Deshabilitar la opción de sólo mostrar los títulos cuando se muestra la vista de búsqueda';
 
   @override
-  String get settings_disable_subdued_note_content_preview => 'Non-subdued preview';
+  String get settings_disable_subdued_note_content_preview => 'No subrayado en vista previa';
 
   @override
   String get settings_disable_subdued_note_content_preview_description =>
-      'Disable the subdued text color of the notes content preview';
+      'Desactiva el color de texto subrayado de la vista previa del contenido de las notas';
 
   @override
-  String get settings_show_tiles_background => 'Background';
+  String get settings_show_tiles_background => 'Fondo';
 
   @override
-  String get settings_show_tiles_background_description => 'Show the background of the notes tiles';
+  String get settings_show_tiles_background_description => 'Muestra el fondo de tarjetas en las notas';
 
   @override
-  String get settings_show_separators => 'Separators';
+  String get settings_show_separators => 'Separadores';
 
   @override
-  String get settings_show_separators_description => 'Show a separator between the notes tiles';
+  String get settings_show_separators_description => 'Mostrar un separador entre las notas';
 
   @override
   String get settings_behavior => 'Comportamiento';
 
   @override
-  String get settings_behavior_description => 'Confirmations, secure flag, swipe actions';
+  String get settings_behavior_description => 'Confirmaciones, marcar cómo segura, acciones de deslizamiento';
 
   @override
-  String get settings_behavior_application => 'Application';
+  String get settings_behavior_application => 'Aplicación';
 
   @override
-  String get settings_confirmations => 'Confirmation dialog';
+  String get settings_confirmations => 'Diálogo de confirmación';
 
   @override
   String get settings_confirmations_description =>
-      'When to show a confirmation dialog when performing an action on a note';
+      'Cuándo mostrar un cuadro de diálogo de confirmación al realizar una acción en una nota';
 
   @override
-  String get settings_confirmations_title_none => 'Never';
+  String get settings_confirmations_title_none => 'Nunca';
 
   @override
-  String get settings_confirmations_title_irreversible => 'Irreversible actions only';
+  String get settings_confirmations_title_irreversible => 'En acciones irreversibles';
 
   @override
-  String get settings_confirmations_title_all => 'Always';
+  String get settings_confirmations_title_all => 'Siempre';
 
   @override
-  String get settings_flag_secure => 'Flag the app as secure';
+  String get settings_flag_secure => 'Marcar la aplicación cómo segura';
 
   @override
   String get settings_flag_secure_description =>
-      'Hide the app from the recent apps and prevent screenshots from being made';
+      'Oculta la aplicación de las aplicaciones recientes y previene que se realicen capturas de pantalla';
 
   @override
-  String get settings_behavior_swipe_actions => 'Swipe actions';
+  String get settings_behavior_swipe_actions => 'Acciones de deslizamiento';
 
   @override
-  String get settings_swipe_action_right => 'Right swipe action';
+  String get settings_swipe_action_right => 'Deslizar a la derecha';
 
   @override
   String get settings_swipe_action_right_description =>
-      'Action to trigger when a right swipe is performed on a note tile';
+      'Acción a realizar al deslizar a la derecha la tarjeta se una nota';
 
   @override
-  String get settings_swipe_action_left => 'Left swipe action';
+  String get settings_swipe_action_left => 'Desliza a la izquierda';
 
   @override
   String get settings_swipe_action_left_description =>
-      'Action to trigger when a left swipe is performed on a note tile';
+      'Acción a realizar al deslizar a la izquierda la tarjeta de una nota';
 
   @override
   String get settings_editor => 'Editor';
 
   @override
-  String get settings_editor_description => 'Buttons, toolbar, reading mode, spacing';
+  String get settings_editor_description => 'Botones, barra de herramientas, modo lectura, espaciado';
 
   @override
-  String get settings_editor_formatting => 'Formatting';
+  String get settings_editor_formatting => 'Formato';
 
   @override
   String get settings_show_undo_redo_buttons => 'Botones deshacer/rehacer';
 
   @override
   String get settings_show_undo_redo_buttons_description =>
-      'Show the buttons to undo and redo changes in the editor\'s app bar';
+      'Mostrar los botones para deshacer y rehacer cambios en la barra de aplicaciones del editor';
 
   @override
   String get settings_show_checklist_button => 'Botón de casilla de verificación';
 
   @override
   String get settings_show_checklist_button_description =>
-      'Show the button to toggle checklists in the editor\'s app bar, hiding it from the editor\'s toolbar if enabled';
+      'Mostrar el botón para alternar listas en la barra de aplicaciones del editor, ocultándolo de la barra de herramientas del editor si está habilitado';
 
   @override
-  String get settings_show_toolbar => 'Toolbar';
+  String get settings_show_toolbar => 'Barra de herramientas';
 
   @override
-  String get settings_show_toolbar_description => 'Show the editor\'s toolbar to enable advanced text formatting';
+  String get settings_show_toolbar_description =>
+      'Mostrar la barra de herramientas del editor para habilitar el formato avanzado de texto';
 
   @override
-  String get settings_editor_behavior => 'Behavior';
+  String get settings_editor_behavior => 'Comportamiento';
 
   @override
-  String get settings_show_editor_mode_button => 'Editor mode button';
+  String get settings_show_editor_mode_button => 'Botón del modo editor';
 
   @override
   String get settings_show_editor_mode_button_description =>
-      'Enable the button to toggle the editor between editing mode and reading mode';
+      'Activar el botón para cambiar el editor entre el modo de edición y el modo de lectura';
 
   @override
-  String get settings_open_editor_reading_mode => 'Open in reading mode';
+  String get settings_open_editor_reading_mode => 'Abrir en modo de lectura';
 
   @override
-  String get settings_open_editor_reading_mode_description => 'Open the editor in reading mode by default';
+  String get settings_open_editor_reading_mode_description => 'Abrir el editor en modo de lectura por defecto';
 
   @override
-  String get settings_focus_title_on_new_note => 'Focus the title';
+  String get settings_focus_title_on_new_note => 'Enfocar el título';
 
   @override
   String get settings_focus_title_on_new_note_description =>
-      'Focus the title instead of the content when creating a new note';
+      'Enfocar el título en lugar del contenido cuando se crea una nueva nota';
 
   @override
-  String get settings_editor_appearance => 'Appearance';
+  String get settings_editor_appearance => 'Apariencia';
 
   @override
-  String get settings_use_paragraph_spacing => 'Paragraph spacing';
+  String get settings_use_paragraph_spacing => 'Espacio entre párrafos';
 
   @override
-  String get settings_use_paragraph_spacing_description => 'Use spacing between paragraphs';
+  String get settings_use_paragraph_spacing_description => 'Usar espaciado entre párrafos';
 
   @override
   String get settings_backup => 'Respaldo';
 
   @override
-  String get settings_backup_description => 'Manual and automatic export, encryption, import';
+  String get settings_backup_description => 'Exportación manual y automática, cifrado, importación';
 
   @override
-  String get settings_backup_import => 'Import';
+  String get settings_backup_import => 'Importar';
 
   @override
   String get settings_import => 'Importar';
@@ -261,109 +288,111 @@ class AppLocalizationsEs extends AppLocalizations {
   String get settings_import_description => 'Importar notas desde un archivo JSON';
 
   @override
-  String get settings_backup_manual_export => 'Manual export';
+  String get settings_backup_manual_export => 'Exportar manualmente';
 
   @override
   String get settings_export_json => 'Exportar a JSON';
 
   @override
   String get settings_export_json_description =>
-      'Immediately export the notes to a JSON file (bin included) that can be imported back';
+      'Exportar inmediatamente las notas a un archivo JSON (bin incluido) que puede ser importado de nuevo';
 
   @override
   String get settings_export_markdown => 'Exportar a Markdown';
 
   @override
-  String get settings_export_markdown_description => 'Immediately export the notes to a Markdown file (bin included)';
+  String get settings_export_markdown_description =>
+      'Exportar inmediatamente las notas a un archivo Markdown (bin incluido)';
 
   @override
-  String get settings_backup_auto_export => 'Automatic export';
+  String get settings_backup_auto_export => 'Exportación automática';
 
   @override
-  String get settings_auto_export => 'Automatic export';
+  String get settings_auto_export => 'Exportación automática';
 
   @override
   String get settings_auto_export_description =>
-      'Automatically export the notes to a JSON file (bin included) that can be imported back';
+      'Exportar automáticamente las notas a un archivo JSON (bin incluido) que puede ser importado de nuevo';
 
   @override
-  String get settings_auto_export_frequency => 'Frequency';
+  String get settings_auto_export_frequency => 'Frecuencia';
 
   @override
   String settings_auto_export_frequency_value(String frequency) {
     String _temp0 = intl.Intl.selectLogic(
       frequency,
       {
-        '1': 'day',
-        '7': 'week',
-        '14': '2 weeks',
-        '30': 'month',
-        'other': '$frequency days',
+        '1': 'día',
+        '7': 'semana',
+        '14': '¡2 semanas',
+        '30': 'mes',
+        'other': '$frequency días',
       },
     );
-    return 'Every $_temp0';
+    return '¡Cada $_temp0';
   }
 
   @override
-  String get settings_auto_export_frequency_description => 'Frequency of the automatic export of the notes';
+  String get settings_auto_export_frequency_description => 'Frecuencia de la exportación automática de las notas';
 
   @override
-  String get settings_auto_export_encryption => 'Encryption';
+  String get settings_auto_export_encryption => 'Encriptación';
 
   @override
   String get settings_auto_export_encryption_description =>
-      'Encrypt the title and the content of the notes with a password';
+      'Encriptar el título y el contenido de las notas con una contraseña';
 
   @override
-  String get settings_auto_export_directory => 'Directory';
+  String get settings_auto_export_directory => 'Directorio';
 
   @override
   String get settings_auto_export_directory_description =>
-      'Directory where to store the automatic exports of the notes';
+      'Directorio donde guardar las exportaciones automáticas de las notas';
 
   @override
   String get settings_about => 'Acerca de';
 
   @override
-  String get settings_about_description => 'Information, help, links';
+  String get settings_about_description => 'Información, ayuda, enlaces';
 
   @override
-  String get settings_about_application => 'Application';
+  String get settings_about_application => 'Aplicación';
 
   @override
-  String get settings_build_mode => 'Build mode';
+  String get settings_build_mode => 'Modo de compilación';
 
   @override
-  String get settings_build_mode_release => 'Release';
+  String get settings_build_mode_release => 'Versión';
 
   @override
-  String get settings_build_mode_debug => 'Debug';
+  String get settings_build_mode_debug => 'Depuración';
 
   @override
-  String get settings_about_help => 'Help';
+  String get settings_about_help => 'Ayuda';
 
   @override
-  String get settings_github_issues => 'Report a bug or request a feature';
+  String get settings_github_issues => 'Reporta errores o solicita nuevas características';
 
   @override
-  String get settings_github_issues_description => 'Report a bug or request a feature by creating a GitHub issue';
+  String get settings_github_issues_description =>
+      'Reportar un error o solicitar una característica creando un problema de GitHub';
 
   @override
-  String get settings_github_discussions => 'Ask a question';
+  String get settings_github_discussions => 'Hacer una pregunta';
 
   @override
-  String get settings_github_discussions_description => 'Ask a question on GitHub discussions';
+  String get settings_github_discussions_description => 'Hacer una pregunta en las discusiones de GitHub';
 
   @override
-  String get settings_get_in_touch => 'Contact the developer';
+  String get settings_get_in_touch => 'Contactar con el desarrollador';
 
   @override
   String settings_get_in_touch_description(Object email) {
-    return 'Contact the developer via mail at $email';
+    return 'Póngase en contacto con el desarrollador por correo electrónico en $email';
   }
 
   @override
-  String get settings_about_links => 'Links';
+  String get settings_about_links => 'Enlaces';
 
   @override
   String get settings_github => 'GitHub';
@@ -375,7 +404,7 @@ class AppLocalizationsEs extends AppLocalizations {
   String get settings_localizations => 'Crowdin';
 
   @override
-  String get settings_localizations_description => 'Add or improve the localizations on the Crowdin project';
+  String get settings_localizations_description => 'Añade o mejora las localizaciones en el proyecto Crowdin';
 
   @override
   String get settings_licence => 'Licencia';
@@ -390,10 +419,10 @@ class AppLocalizationsEs extends AppLocalizations {
   String get hint_note => 'Nota';
 
   @override
-  String get hint_link => 'Link';
+  String get hint_link => 'Enlace';
 
   @override
-  String get dialog_export_encryption_password => 'Password';
+  String get dialog_export_encryption_password => 'Contraseña';
 
   @override
   String get tooltip_toggle_checkbox => 'Alternar casilla';
@@ -408,10 +437,10 @@ class AppLocalizationsEs extends AppLocalizations {
   String get tooltip_fab_empty_bin => 'Vaciar la papelera';
 
   @override
-  String get tooltip_fab_toggle_editor_mode_edit => 'Switch to editing mode';
+  String get tooltip_fab_toggle_editor_mode_edit => 'Cambiar al modo de edición';
 
   @override
-  String get tooltip_fab_toggle_editor_mode_read => 'Switch to reading mode';
+  String get tooltip_fab_toggle_editor_mode_read => 'Cambiar al modo de lectura';
 
   @override
   String get tooltip_layout_list => 'Vista en lista';
@@ -441,10 +470,10 @@ class AppLocalizationsEs extends AppLocalizations {
   String get tooltip_restore => 'Restaurar';
 
   @override
-  String get tooltip_reset => 'Reset';
+  String get tooltip_reset => 'Restablecer';
 
   @override
-  String get dialog_add_link => 'Add a link';
+  String get dialog_add_link => 'Agregar un vínculo';
 
   @override
   String get dialog_delete => 'Eliminar';
@@ -513,66 +542,66 @@ class AppLocalizationsEs extends AppLocalizations {
       '¿Realmente quieres restaurar la papelera? No podrás restaurar las notas en ella.';
 
   @override
-  String get dialog_export_encryption_switch => 'Encrypt the JSON export';
+  String get dialog_export_encryption_switch => 'Encriptar la exportación JSON';
 
   @override
   String get dialog_export_encryption_description =>
-      'The title and the content of the notes will be encrypted using your password. It should be randomly generated, exactly 32 characters long, strong (at least 1 lowercase, 1 uppercase, 1 number and 1 special character) and securely stored.';
+      'El título y el contenido de las notas serán cifrados usando su contraseña. Debe ser generado aleatoriamente, exactamente 32 caracteres de largo, fuerte (al menos 1 minúscula, 1 mayúscula, 1 número y 1 carácter especial) y almacenado de forma segura.';
 
   @override
   String get dialog_export_encryption_secondary_description_auto =>
-      'This password will be used for all future automatic exports.';
+      'Esta contraseña se utilizará para todas las exportaciones automáticas futuras.';
 
   @override
   String get dialog_export_encryption_secondary_description_manual =>
-      'This password will only be used for this export.';
+      'Esta contraseña sólo se usará para esta exportación.';
 
   @override
-  String get dialog_export_encryption_password_invalid => 'Invalid';
+  String get dialog_export_encryption_password_invalid => 'Inválido';
 
   @override
   String get dialog_import_encryption_password_description =>
-      'This export is encrypted. To import it, you need to provide the password used to encrypt it.';
+      'Esta exportación está encriptada. Para importarla, necesita proporcionar la contraseña utilizada para encriptarla.';
 
   @override
   String get dialog_import_encryption_password_error =>
-      'the decrypting of the export failed. Please check that you provided the same password that the one you used for encrypting the export.';
+      'el desencriptado de la exportación falló. Por favor, compruebe que proporcionó la misma contraseña que la que utilizó para encriptar la exportación.';
 
   @override
-  String get button_sort_date => 'Date';
+  String get button_sort_date => 'Fecha';
 
   @override
   String get placeholder_notes => 'No hay notas';
 
   @override
-  String get placeholder_bin => 'No deleted notes';
+  String get placeholder_bin => 'No hay notas eliminadas';
 
   @override
-  String get action_disabled => 'Disabled';
+  String get action_disabled => 'Desactivado';
 
   @override
-  String get action_pin => 'Pin';
+  String get action_pin => 'Fijar';
 
   @override
-  String get action_unpin => 'Unpin';
+  String get action_unpin => 'Desfijar';
 
   @override
-  String get action_copy => 'Copy';
+  String get action_copy => 'Copiar';
 
   @override
-  String get action_share => 'Share';
+  String get action_share => 'Compartir';
 
   @override
-  String get action_delete => 'Delete';
+  String get action_delete => 'Eliminar';
 
   @override
-  String get action_restore => 'Restore';
+  String get action_restore => 'Restaurar';
 
   @override
-  String get action_delete_permanently => 'Delete permanently';
+  String get action_delete_permanently => 'Eliminar permanentemente';
 
   @override
-  String get action_about => 'About';
+  String get action_about => 'Acerca de';
 
   @override
   String get about_last_edited => 'Última edición';
@@ -587,16 +616,22 @@ class AppLocalizationsEs extends AppLocalizations {
   String get about_characters => 'Caracteres';
 
   @override
-  String get about_time_at => 'at';
+  String get about_time_at => 'a las';
 
   @override
-  String get snack_bar_copied => 'Content of the note copied to the clipboard.';
+  String get snack_bar_copied => 'Contenido de la nota copiada al portapapeles.';
 
   @override
-  String get snack_bar_import_success => 'The notes were successfully imported.';
+  String get snack_bar_import_success => 'Las notas fueron importadas exitosamente.';
 
   @override
-  String get snack_bar_export_success => 'The notes were successfully exported.';
+  String get snack_bar_export_success => 'Las notas fueron exportadas exitosamente.';
+
+  @override
+  String get snack_bar_logs_copied => 'The logs were copied to your clipboard.';
+
+  @override
+  String get snack_bar_logs_exported => 'The logs were successfully exported.';
 
   @override
   String get action_add_note_title => 'Agregar una nota';