+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.7.0 - 2024-10-..
+## 1.7.1 - 2024-10-11
+
+### Added
+
+- Error page shown when a page fails to load, with the ability to copy and export the error logs
+
+### Fixed
+
+- Value of the setting on the settings tiles not using a subdued color when disabled
+- Small note pinned icon on the notes tiles
+- Small application icon
+
+## 1.7.0 - 2024-10-05
 
 ### Added